Fix the default image provider.
